H2O: Move the test implementations to a separate directory
