+++ frameworks/C/h2o/src/handlers/fortune.h
