+++ frameworks/C/h2o/src/handlers/json_serializer.c
